requirements.txt
backports.functools-lru-cache==1.5
beautifulsoup4==4.8.0
boto==2.49.0
boto3==1.9.214
botocore==1.12.214
bs4==0.0.1
bz2file==0.98
certifi==2019.6.16
chardet==3.0.4
Click==7.0
decorator==4.4.0
Django==1.11
django-bootstrap4==0.0.8
djangorestframework==3.9.4
docutils==0.15.2
enum34==1.1.6
futures==3.3.0
gensim==3.8.0
gTTS==2.0.3
gTTS-token==1.1.3
idna==2.8
imageio==2.4.1
jmespath==0.9.4
moviepy==1.0.0
nltk==3.4.5
numpy==1.16.1
pdfminer==20140328
Pillow==6.1.0
proglog==0.1.9
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
s3transfer==0.2.1
scipy==1.2.2
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.8.4
soupsieve==1.9.3
tqdm==4.34.0
urllib3==1.25.3